+++ packages/components/src/border-box-control/types.ts
@@ -14,14 +14,6 @@ export type AnyBorder = Border | Borders | undefined;
 export type BorderProp = keyof Border;
 export type BorderSide = keyof Borders;
 
-export type PopoverClassNames = {
-	linked?: string;
-	top?: string;
-	right?: string;
-	bottom?: string;
-	left?: string;
-};
-
 export type BorderBoxControlProps = ColorProps &
 	LabelProps & {
 		/**
@@ -35,10 +27,13 @@ export type BorderBoxControlProps = ColorProps &
 		 */
 		onChange: ( value: AnyBorder ) => void;
 		/**
-		 * An object defining CSS classnames for all the inner `BorderControl`
-		 * popover content.
+		 * The position of the color popovers compared to the control wrapper.
 		 */
-		popoverClassNames?: PopoverClassNames;
+		popoverPlacement?: string;
+		/**
+		 * The space between the popover and the control wrapper.
+		 */
+		popoverOffset?: number;
 		/**
 		 * An object representing the current border configuration.
 		 *
@@ -106,10 +101,13 @@ export type SplitControlsProps = ColorProps & {
 	 */
 	onChange: ( value: Border | undefined, side: BorderSide ) => void;
 	/**
-	 * An object defining CSS classnames for the split side `BorderControl`s'
-	 * popover content.
+	 * The position of the color popovers compared to the control wrapper.
 	 */
-	popoverClassNames?: PopoverClassNames;
+	popoverPlacement?: string;
+	/**
+	 * The space between the popover and the control wrapper.
+	 */
+	popoverOffset?: number;
 	/**
 	 * An object representing the current border configuration. It contains
 	 * properties for each side, with each side an object reflecting the border